+++ .gitignore
@@ -32,14 +32,6 @@ bld/
 [Ll]og/
 [Ll]ogs/
 
-# Ignore VSCode
-.vscode/*
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
-!.vscode/*.code-snippets
-
 # Local History for Visual Studio Code
 .history/
