--- .vscode/launch.json
@@ -1,21 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": ".NET Core Launch (console)",
-            "type": "coreclr",
-            "request": "launch",
-            "preLaunchTask": "build",
-            "program": "${workspaceFolder}/EchoServer/bin/Debug/net6.0/EchoServer.dll",
-            "args": [],
-            "cwd": "${workspaceFolder}/EchoServer",
-            "console": "internalConsole",
-            "stopAtEntry": false
-        },
-        {
-            "name": ".NET Core Attach",
-            "type": "coreclr",
-            "request": "attach"
-        }
-    ]
-}
