--- TKeazirian.HTTPServer.Tests/UnitTest1.cs
@@ -1,11 +1,0 @@
-using Xunit;
-
-namespace TKeazirian.HTTPServer.Tests;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1()
-    {
-    }
-}
