+++ TKeazirian.HTTPServer/Program.cs
@@ -1,3 +1,11 @@
-﻿// See https://aka.ms/new-console-template for more information
-
-Console.WriteLine("Hello, World!");
+﻿namespace TKeazirian.HTTPServer;
+public static class Program
+{
+    private static void Main(string[] args)
+    {
+        var g = new Greeting();
+        var hour = DateTime.Now.Hour;
+        var greeting = Greeting.GetGreeting(hour);
+        Console.WriteLine(greeting);
+    }
+}
